requirements.txt
yfinance==0.2.12
pandas==2.0.3
numpy==1.23.5
alpaca-trade-api==2.3.0
python-dotenv==1.0.0
